requirements.txt
streamlit
langchain
langgraph
langchain_openai
langchain_core
msal
python-dotenv
asyncio
